intelligence/geoint-engine/surveillance-networks/requirements.txt
# Apollo Platform - Facial/Audio Recognition Requirements
# Agent 5 - Elite Level Dependencies

# Core Facial Recognition
face-recognition==1.3.0
dlib==19.24.2  # Face recognition backend
cmake==3.28.1  # Required for dlib compilation

# Computer Vision
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
Pillow==10.2.0

# Image Processing
scikit-image==0.22.0
imageio==2.33.1

# Audio Processing
openai-whisper==20231117  # Speech recognition
SpeechBrain==0.5.16  # Voice biometrics
librosa==0.10.1  # Audio analysis
soundfile==0.12.1
audioread==3.0.1

# Machine Learning & AI
torch==2.1.2  # PyTorch for deep learning
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.36.2  # Hugging Face models
accelerate==0.25.0

# Age Progression & Enhancement
insightface==0.7.3  # Advanced face analysis
onnxruntime==1.16.3
mxnet==1.9.1

# Data Processing
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# HTTP & API
requests==2.31.0
aiohttp==3.9.1

# Utilities
tqdm==4.66.1  # Progress bars
python-dotenv==1.0.0
pyyaml==6.0.1

# Database Connectivity (for storing results)
psycopg2-binary==2.9.9  # PostgreSQL
redis==5.0.1  # Redis
neo4j==5.15.0  # Neo4j graph database
